docs: Rearrange the folder structure for better presentation
The ``inputFiles.rst`` used to describe all things related to the ``parameters.in``. Since we include all kinds of input files and present them in the Folder Structure section, the ``inputFiles.rst`` is now the introduction page to all other input files.
The descriptions of the format of the 3D Velocity Model and the ``topography.in`` are now independent files.
A simple and under-constructed Source Input File section is added.
All other ``parameters.in``-related pages are moved to the ``parameters`` folder under the ``input`` folder.