docs: Add plotmesh.m section
The MATLAB script used to plot mesh data, `plotmesh.m`, has been fixed in the new release version of Hercules.
A new section is added to the documentation accordingly.
In addition, some explanation has also been added to the Compiling section to make it easier for users to understand how to set up a `user.mk`.